Refactor column name formatter
