+++ quesma/clickhouse/util.go
@@ -12,6 +12,25 @@ import (
 	"time"
 )
 
+type TableColumNameFormatter interface {
+	Format(namespace, columnName string) string
+}
+
+type columNameFormatter struct {
+	separator string
+}
+
+func (t *columNameFormatter) Format(namespace, columnName string) string {
+	if namespace == "" {
+		return columnName
+	}
+	return fmt.Sprintf("%s%s%s", namespace, t.separator, columnName)
+}
+
+func DefaultColumnNameFormatter() TableColumNameFormatter {
+	return &columNameFormatter{separator: "::"}
+}
+
 // Code doesn't need to be pretty, 99.9% it's just for our purposes
 // Parses type from SHOW COLUMNS FROM "table"
 func parseTypeFromShowColumns(typ, name string) (Type, string) {
