+++ quesma/plugins/registry/plugin_registry.go
@@ -3,7 +3,6 @@
 package registry
 
 import (
-	"fmt"
 	"quesma/plugins"
 	"quesma/quesma/config"
 	"quesma/schema"
@@ -12,7 +11,6 @@ import (
 // Plugin changes the behavior of Quesma by changing the pipeline of transformers
 type Plugin interface {
 	ApplyResultTransformers(table string, cfg config.QuesmaConfiguration, schema schema.Registry, transformers []plugins.ResultTransformer) []plugins.ResultTransformer
-	GetTableColumnFormatter(table string, cfg config.QuesmaConfiguration, schema schema.Registry) plugins.TableColumNameFormatter
 }
 
 var registeredPlugins []Plugin
@@ -39,25 +37,3 @@ func ResultTransformerFor(table string, cfg config.QuesmaConfiguration, schema s
 }
 
 ///
-
-func TableColumNameFormatterFor(table string, cfg config.QuesmaConfiguration, schema schema.Registry) (plugins.TableColumNameFormatter, error) {
-
-	var transformers []plugins.TableColumNameFormatter
-
-	for _, plugin := range registeredPlugins {
-		t := plugin.GetTableColumnFormatter(table, cfg, schema)
-		if t != nil {
-			transformers = append(transformers, t)
-		}
-	}
-
-	if len(transformers) == 0 {
-		return nil, fmt.Errorf("no table column name formatter found for table %s", table)
-	}
-
-	if len(transformers) > 1 {
-		return nil, fmt.Errorf("multiple table column name formatters are not supported, table %s", table)
-	}
-
-	return transformers[0], nil
-}
