+++ quesma/quesma/functionality/bulk/bulk.go
@@ -12,7 +12,6 @@ import (
 	"quesma/clickhouse"
 	"quesma/jsonprocessor"
 	"quesma/logger"
-	"quesma/plugins/registry"
 	"quesma/queryparser"
 	"quesma/quesma/config"
 	"quesma/quesma/recovery"
@@ -204,11 +203,9 @@ func sendToClickhouse(ctx context.Context, clickhouseDocumentsToInsert map[strin
 				inserts[i] = document.document
 			}
 
-			nameFormatter, err := registry.TableColumNameFormatterFor(indexName, cfg, nil)
-			if err == nil {
-				transformer := jsonprocessor.IngestTransformerFor(indexName, cfg)
-				err = lm.ProcessInsertQuery(ctx, indexName, inserts, transformer, nameFormatter)
-			}
+			nameFormatter := clickhouse.DefaultColumnNameFormatter()
+			transformer := jsonprocessor.IngestTransformerFor(indexName, cfg)
+			err := lm.ProcessInsertQuery(ctx, indexName, inserts, transformer, nameFormatter)
 
 			for _, document := range documents {
 				bulkSingleResponse := BulkSingleResponse{
